Revert "Merge pull request #2018 from planetarium/improve/has-execute"
This reverts commit f76932c200a7766cedb7e15d4c40e4bdded0c504, reversing
changes made to ddb70d8e160030a8f3cdb8ba264f3b7ba69a6bc2.

+++ Lib9c/Model/Character/ArenaCharacter.cs
@@ -819,7 +819,7 @@ namespace Nekoyume.Model
                 return CRI >= chance;
 
             var additionalCriticalChance =
-                AttackCountHelper.GetAdditionalCriticalChance(_attackCount, _attackCountMax);
+                (int) AttackCountHelper.GetAdditionalCriticalChance(_attackCount, _attackCountMax);
             return CRI + additionalCriticalChance >= chance;
         }
 
@@ -850,7 +850,7 @@ namespace Nekoyume.Model
                 _attackCount = 1;
             }
 
-            var damageMultiplier = AttackCountHelper.GetDamageMultiplier(_attackCount, _attackCountMax);
+            var damageMultiplier = (int) AttackCountHelper.GetDamageMultiplier(_attackCount, _attackCountMax);
             damage *= damageMultiplier;
             return damage;
         }
